tests/wpt/web-platform-tests/sanitizer-api/sanitizer-insecure-context.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
  // Currently, the Sanitizer requires a secure context.
  test(t => {
    assert_false(globalThis.isSecureContext);
    assert_equals("Sanitizer" in globalThis, globalThis.isSecureContext);
    assert_equals("setHTML" in document.body, globalThis.isSecureContext);
  }, "Sanitizer API in an insecure context.");
</script>
</body>
</html>
