third_party/blink/web_tests/virtual/sanitizer-api/external/wpt/sanitizer-api/sanitizer-insecure-context-expected.txt

This is a testharness.js-based test.
[FAIL] Sanitizer API in an insecure context.
  assert_equals: expected false but got true
Harness: the test ran to completion.
